ABBCK-39 # Dodanie wyswietlenia informacji o pracowniku i dodanie klienta do bazy

+++ src/main/java/hibernate/entity/Klienci.java
@@ -29,7 +29,7 @@ public class Klienci {
     public Klienci() {
     }
 
-    public long getIdKlienta() {
+    public int getIdKlienta() {
         return idKlienta;
     }
